Refactor: 사진을 처음 등록할 때 null 사진 키 값으로 예외 발생으로 인해 존재할 경우에만 삭제하도록 로직 추가

+++ src/main/java/project/closet/user/service/basic/BasicUserService.java
@@ -100,7 +100,10 @@ public class BasicUserService implements UserService {
         user.updateProfile(profileUpdateRequest);
         Optional.ofNullable(profileImage)
                 .map(image -> {
-                    s3ContentStorage.deleteByKey(user.getProfile().getProfileImageKey());
+                    if (user.getProfile().getProfileImageKey() != null) {
+                        log.debug("기존 프로필 이미지 삭제: {}", user.getProfile().getProfileImageKey());
+                        s3ContentStorage.deleteByKey(user.getProfile().getProfileImageKey());
+                    }
                     return s3ContentStorage.upload(image);
                 })
                 .ifPresent(user::updateProfileImageKey);
